requirements.txt
streamlit
pandas
textblob
Pillow
matplotlib
openpyxl
PyMuPDF
